Extract EXCEPTION_LOCATION
Extract from THROW macro part that is responsible for injecting
information about the location (file/line/function) where the exception
occurred.

Remarks: this solution isn't probably the best idea but don't have time
now for thinking about a better when; however, keep in mind that it
needs to be refactored.