Drop storing task result in aligned storage.
Task result, disregarding its type, was stores in a aligned storage.
Since the type can be arbitrarilly large this solution was dropped in
favour of storing the result on the heap for anything else than integral
values.

TODO: Better approach would probably be to store the result in a aligned
storage up to some threshhol value of the size of the result value.